requirements.txt
amqp==5.0.2
asgiref==3.3.1
attrs==20.3.0
autopep8==1.5.4
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cloudpickle==1.6.0
croniter==0.3.36
dask==2020.12.0
distributed==2020.12.0
dj-database-url==0.5.0
Django==3.1.4
django-celery-beat==2.1.0
django-celery-results==2.0.0
django-cors-headers==3.6.0
django-filter==2.4.0
django-pagedown==2.2.0
django-timezone-field==4.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
docker==4.4.0
gunicorn==20.0.4
HeapDict==1.0.1
idna==2.10
importlib-resources==3.3.0
kombu==5.0.2
marshmallow==3.10.0
marshmallow-oneofschema==2.1.0
more-itertools==8.6.0
msgpack==1.0.2
mypy-extensions==0.4.3
natsort==7.1.0
numpy==1.19.4
packaging==20.8
pandas==1.1.5
pendulum==2.1.2
Pillow==8.0.1
pluggy==0.13.1
prefect==0.14.0
prompt-toolkit==3.0.8
psutil==5.8.0
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.6.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.3
python-box==5.2.0
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-slugify==4.0.1
pytz==2020.4
pytzdata==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.25.1
six==1.15.0
sortedcontainers==2.3.0
sqlparse==0.4.1
tabulate==0.8.7
tblib==1.7.0
text-unidecode==1.3
toml==0.10.2
toolz==0.11.1
tornado==6.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
websocket-client==0.57.0
whitenoise==5.2.0
zict==2.0.0
